Reinstate skipping IDS redeploy if props unchanged

+++ src/test/scripts/prepare_test.py
@@ -30,13 +30,10 @@ try:
             print(t, end="", file=f)
         print("icat.url = %s" % icaturl, file=f)
         print("testHome = %s" % home, file=f)
-    # TODO: sort this out
-    # presumably meant to prevent reinstalling of the IDS
-    # when run.properties has not changed but also stops
-    # setup install being run under other conditions        
-#    if (os.path.exists("src/test/install/run.properties") and 
-#        cmp(name, "src/test/install/run.properties")):
-#        sys.exit(0)
+    if (os.path.exists("src/test/install/run.properties") and 
+        cmp(name, "src/test/install/run.properties")):
+        print("run.properties has not changed - skipping IDS redeploy")
+        sys.exit(0)
     print("Installing with %s" % propFile)
     shutil.copy(name, "src/test/install/run.properties")
 finally:
